Remove garage.misc.console
* Remove module garage.misc.console
* Remove a few minor usages of colorize
* Move query_yes_no to the only example which uses it
* Delete some dead code from garage.experiment.experiment
